+++ frontend/src/components/Gallery/Gallery.jsx
@@ -3,7 +3,7 @@ import { IMAGE_TYPE_GIFS } from "../../static/constants";
 
 import Modal from "react-bootstrap/Modal";
 
-const SearchViewer = ({
+const Gallery = ({
   images,
   imageType,
 }) => {
@@ -73,4 +73,4 @@ const SearchViewer = ({
   );
 };
 
-export default SearchViewer;
+export default Gallery;
