+++ frontend/src/components/LoadMoreButton/LoadMoreButton.jsx
@@ -1,11 +1,11 @@
 import React, { useState } from "react";
 import { LIMIT } from "../../static/constants";
 import { IMAGE_TYPE_GIFS } from "../../static/constants";
-import { getImages, getSearchedImages } from "../../utils/getImage";
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faAngleDown } from "@fortawesome/free-solid-svg-icons";
 import Loading from "../Loading/Loading";
+import getImages from "../../utils/getImages";
 
 const LoadMoreButton = ({
   imageType,
@@ -16,32 +16,20 @@ const LoadMoreButton = ({
 }) => {
   const [isLoading, setIsLoading] = useState(false);
 
-  const updateOffsetValue = () => {
-    imageType === IMAGE_TYPE_GIFS
-      ? setOffset(offset + LIMIT + 1)
-      : setOffset(offset + 1);
-  };
+  function getNewOffset(prevOffset) {
+    return imageType === IMAGE_TYPE_GIFS ? prevOffset + LIMIT + 1 : offset + 1;
+  }
 
   const loadMoreImages = async () => {
     setIsLoading(true);
 
-    try {
-      const receivedImages = !inputValue
-        ? await getImages(imageType, LIMIT, offset)
-        : await getSearchedImages(
-          imageType,
-          inputValue.replace(/[^a-zA-Z ]/g, ""),
-          LIMIT,
-          offset
-        );
-
-      setImages((images) => [...images, ...receivedImages]);
-      updateOffsetValue();
-    } catch (error) {
-      console.error(error);
-    } finally {
-      setIsLoading(false);
-    }
+    getImages(imageType, offset, inputValue)
+      .then(receivedImages => {
+        const newOffset = getNewOffset(offset);
+        setImages((images) => [...images, ...receivedImages]);
+        setOffset(newOffset);
+        setIsLoading(false);
+      })
   };
 
   return (
