+++ frontend/src/components/LoadMoreButton/_LoadMoreButton.scss
@@ -3,6 +3,7 @@
   justify-content: center;
   align-items: center;
   margin: 20px;
+  height: 32px;
 }
 
 .moreImagesButton {
