+++ frontend/src/utils/getImages.js
@@ -1,11 +1,11 @@
-import { IMAGE_TYPE_GIFS } from "../static/constants";
+import { IMAGE_TYPE_GIFS, LIMIT } from "../static/constants";
 
 const axios = require("axios");
 const hostName = `${
   process?.env?.REACT_APP_API_URL || "http://localhost:5000"
 }`;
 
-export const getImages = async (imageType, limit, offset) => {
+const fetchImages = async (imageType, limit, offset) => {
   const response = await axios.get(
     `${hostName}/${imageType}?limit=${limit}&offset=${offset}`
   );
@@ -14,7 +14,7 @@ export const getImages = async (imageType, limit, offset) => {
   return response.data[propertyName];
 };
 
-export const getSearchedImages = async (
+const fetchSearchedImages = async (
   imageType,
   searchTerm,
   limit,
@@ -27,3 +27,24 @@ export const getSearchedImages = async (
 
   return response.data[propertyName];
 };
+
+const getImages = async (imageType, offsetToFetch, searchTerm) => {
+  try {
+    if (searchTerm) {
+      const formattedSearchTerm = searchTerm.replace(/[^a-zA-Z ]/g, "");
+
+      return await fetchSearchedImages(
+        imageType,
+        formattedSearchTerm,
+        LIMIT,
+        offsetToFetch
+      );
+    } else {
+      return await fetchImages(imageType, LIMIT, offsetToFetch);
+    }
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+export default getImages;
